add deploy script presets

+++ contracts/Web3Workshop.sol
@@ -9,13 +9,12 @@ import "./NFTTicket.sol";
 /// @dev Unaudited contract for learning purposes only
 //  Note: contract is equivalent to class keyword
 contract Web3Workshop {
-
-  // STATE VARIABLES 
+  // STATE VARIABLES
   // State variables live on the blockchain and act like storage
   // The more state variables the more the contract costs to deploy to chain
-  uint public ticketsSold;
-  uint public maxAttendees;
-  uint public basePrice;
+  uint256 public ticketsSold;
+  uint256 public maxAttendees;
+  uint256 public basePrice;
 
   // Address are primitive types in Ehereum
   // Think of them like your email or bank number
@@ -25,31 +24,31 @@ contract Web3Workshop {
   // Create NFT ticket contract instance
   NFTTicket nftTicket = NFTTicket(owner);
 
-  // Do we want a bool for if someone is attending, 
-  // or rather how many tickets they have, 
+  // Do we want a bool for if someone is attending,
+  // or rather how many tickets they have,
   // so can that be an array?
 
   // Used to create a custom data type
   struct Attendent {
     address attendent;
-    uint ticketQuantity;
-    uint[] tickets;
+    uint256 ticketQuantity;
+    uint256[] tickets;
   }
 
   // public keyword creates an automatic getter
-  // mapping is equivalent to JS Object, a key-value pair / hash map 
+  // mapping is equivalent to JS Object, a key-value pair / hash map
   // see: https://solidity-by-example.org/mapping/
   /// @notice holds public state of the attendee list
   mapping(address => Attendent) public attendees;
 
   NFTTicket nftticket = NFTTicket(owner);
 
-  /// EVENTS 
+  /// EVENTS
   // Events alert the front end to what happened
-  event TicketBought(uint tickets);
+  event TicketBought(uint256 tickets);
 
   /// @notice This constructor sets up the max amount of event attendees
-  constructor(uint _maxAttendees, uint _basePrice) {
+  constructor(uint256 _maxAttendees, uint256 _basePrice) {
     ticketsSold = 0;
     maxAttendees = _maxAttendees;
 
@@ -66,35 +65,44 @@ contract Web3Workshop {
     // require conditional to be true, or reverts back to previous state
     // gas fee to pay for computation is still consumed.
     // see: https://solidity-by-example.org/gas/
-    // see: https://www.geeksforgeeks.org/solidity-error-handling/ 
+    // see: https://www.geeksforgeeks.org/solidity-error-handling/
     require(ticketsSold < maxAttendees, "All sold out.");
-    // indicates to call function. 
+    // indicates to call function.
     // Can be placed before or after code and used multiple times
-    _; 
+    _;
   }
 
   /// @notice calculate and check if funds send match total ticket price
-  modifier checkPayment(uint _amountOfTickets) {
-    require(_amountOfTickets * basePrice <= msg.value, "Payment sent is less than total ticket price");
+  modifier checkPayment(uint256 _amountOfTickets) {
+    require(
+      _amountOfTickets * basePrice <= msg.value,
+      "Payment sent is less than total ticket price"
+    );
     _;
   }
 
   /// @notice allows user to buy ticket if not sold out
   /// first checks if tickets are not sold out - canAttend
   /// then checks total payment matches funds sent - checkPayment
-  function buyTicket(uint _amountOfTickets) external payable canAttend checkPayment(_amountOfTickets) returns (bool) {
+  function buyTicket(uint256 _amountOfTickets)
+    external
+    payable
+    canAttend
+    checkPayment(_amountOfTickets)
+    returns (bool)
+  {
     // Pay contract for ticket
     payable(owner).transfer(msg.value);
 
-    // Add ticket owner to mapping which holds our smart contract's "database"  
+    // Add ticket owner to mapping which holds our smart contract's "database"
     attendees[msg.sender].attendent = msg.sender;
     attendees[msg.sender].ticketQuantity = _amountOfTickets;
-    
+
     // mint number of tickets bought
     buyLoop(_amountOfTickets);
 
     // Increase tickets sold count by num of tickets.
-    ticketsSold = ticketsSold +  _amountOfTickets;
+    ticketsSold = ticketsSold + _amountOfTickets;
 
     // Let front end know of tickets
     emit TicketBought(_amountOfTickets);
@@ -105,14 +113,10 @@ contract Web3Workshop {
 
   /// @notice buys number of tickets ordered
   /// @dev not very efficient, could be better ways of doing this
-  function buyLoop(uint _ticketsBought) internal {
-    for (uint i = 0; i < _ticketsBought; i++) {
+  function buyLoop(uint256 _ticketsBought) internal {
+    for (uint256 i = 0; i < _ticketsBought; i++) {
       // mint NFT to call
-      attendees[msg.sender].tickets.push(
-        nftticket.mintNFT(msg.sender)
-      );
+      attendees[msg.sender].tickets.push(nftticket.mintNFT(msg.sender));
     }
   }
-
-
 }
